assistant2: Factor out tool use into its own module
